pyproject.toml
[project]
name = "meow-bank"
version = "0.1.0"
description = "Bank API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "sqlalchemy>=2.0.41",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.2",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=3.6.0",
    "ruff>=0.11.9",
]

[tool.black]
line-length = 88
force-exclude = '''
/(
    .venv/
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip = [".venv"]

[tool.ruff]
lint.select = ["E", "F", "I", "B", "C4", "UP", "N", "W", "PL", "RUF"]
lint.fixable = ["ALL"]
line-length = 88
exclude = [".venv"]
lint.ignore = ["B008", "RUF010"]
